+++ controller/courses.js
@@ -45,10 +45,15 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 
 exports.addCourse = asyncHandler(async (req, res, next) => {
     req.body.bootcamp = req.params.bootcampId
+    req.body.user = req.user.id
     const bootcamp = await Bootcamp.findById(req.params.bootcampId)
     if (!bootcamp) {
         return next(
             new ErrorResponse(`bootcamp not found with the id of ${req.params.bootcampId}`, 404))
+    }
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`the user with id ${req.user.id} is not authorized to add a course with the bootcamp of id ${req.params.bootcampId}`, 401))
     }
     const course = await Course.create(req.body)
     res.status(200).json({ success: true, data: course });
@@ -64,6 +69,10 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
     if (!course) {
         return next(
             new ErrorResponse(`course not found with the id of ${req.params.id}`, 404))
+    }
+    if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`the user with id ${req.user.id} is not authorized to update the course with the course of id ${course._id}`, 401))
     }
     course = await Course.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
@@ -81,6 +90,10 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
     if (!course) {
         return next(
             new ErrorResponse(`course not found with the id of ${req.params.id}`, 404))
+    }
+    if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`the user with id ${req.user.id} is not authorized to delete the course with the course of id ${course._id}`, 401))
     }
     await Course.findByIdAndDelete(req.params.id)
     res.status(200).json({ success: true, data: {} });
